Time: 6 ms (3.43%), Space: 66.7 MB (7.38%) - LeetHub

+++ LeetCode/Medium/2270-number-of-ways-to-split-array/2270-number-of-ways-to-split-array.java
@@ -1,7 +1,7 @@
 class Solution {
 
     static int n;
-    static int[] pSum, pSumReversed;
+    static long[] pSum, pSumReversed;
 
     public int waysToSplitArray(int[] nums) {
         init(nums);
@@ -33,7 +33,7 @@ class Solution {
 
     static void init(int[] nums) {
         n = nums.length;
-        pSum = new int[n];
-        pSumReversed = new int[n];
+        pSum = new long[n];
+        pSumReversed = new long[n];
     }
 }
